Renaming the OrmCache to SchemaCache
