+++ tests/TestCase/Database/SchemaCacheTest.php
